+++ BeeHive.L30.Domain/SL10.IRepository/Common/IEmployeeRepository.cs
@@ -6,9 +6,9 @@ namespace BeeHive.L30.Domain.SL10.IRepository.Common
     public interface IEmployeeRepository
     {
         List<Employees> All();
-        Employees GetById(int id);
+        Employees GetById(long id);
         Employees Create(Employees domain);
         Employees Update(Employees domain);
-        void Delete(int id);
+        void Delete(long id);
     }
 }
