+++ BeeHive.L30.Domain/SL10.IRepository/Common/IMenuRepository.cs
@@ -7,10 +7,10 @@ namespace BeeHive.L30.Domain.SL10.IRepository.Common
     public interface IMenuRepository
     {
         List<Menu> All();
-        Menu GetById(int id);
+        Menu GetById(long id);
         Menu Create(Menu domain);
         Menu Update(Menu domain);
-        void Delete(int id);
+        void Delete(long id);
         List<MenuLookup> GetAllMenuLookup();
     }
 }
