+++ BeeHive.L30.Domain/SL10.IRepository/Common/IRoleMenuRepository.cs
@@ -6,9 +6,9 @@ namespace BeeHive.L30.Domain.SL10.IRepository.Common
     public interface IRoleMenuRepository
     {
         List<RoleMenu> All();
-        RoleMenu GetById(int id);
+        RoleMenu GetById(long id);
         RoleMenu Create(RoleMenu domain);
         RoleMenu Update(RoleMenu domain);
-        void Delete(int id);
+        void Delete(long id);
     }
 }
