+++ BeeHive.L30.Domain/SL10.IRepository/Common/IRoleRepository.cs
@@ -6,9 +6,9 @@ namespace BeeHive.L30.Domain.SL10.IRepository.Common
     public interface IRoleRepository
     {
         List<Role> All();
-        Role GetById(int id);
+        Role GetById(long id);
         Role Create(Role domain);
         Role Update(Role domain);
-        void Delete(int id);
+        void Delete(long id);
     }
 }
