+++ BeeHive.L30.Domain/SL10.IRepository/Common/IUserRepository.cs
@@ -8,10 +8,10 @@ namespace BeeHive.L30.Domain.SL10.IRepository.Common
     public interface IUserRepository
     {
         List<Hopper> All();
-        Hopper GetById(int id);
+        Hopper GetById(long id);
         Hopper Create(Hopper domain);
         Hopper Update(Hopper domain);
-        void Delete(int id);
+        void Delete(long id);
         Hopper GetUserByUserNameAndPassword(string username,string password);
     }
 }
